backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
ortools>=9.15.6755
googlemaps>=4.10.0
pytesseract>=0.3.10
Pillow>=10.4.0
google-cloud-speech>=2.24.1
httpx>=0.26.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
